requirements.txt
streamlit
PyPDF2
python-docx
openai
scikit-learn
python-dotenv
